+++ desktop_version/src/main.cpp
@@ -31,9 +31,12 @@
 
 scriptclass script;
 
-#if !defined(NO_CUSTOM_LEVELS)
+#ifndef NO_CUSTOM_LEVELS
 std::vector<edentities> edentity;
+customlevelclass cl;
+# ifndef NO_EDITOR
 editorclass ed;
+# endif
 #endif
 
 UtilityClass help;
@@ -620,14 +623,14 @@ int main(int argc, char *argv[])
         game.menustart = true;
 
         LevelMetaData meta;
-        if (ed.getLevelMetaData(playtestname, meta)) {
-            ed.ListOfMetaData.clear();
-            ed.ListOfMetaData.push_back(meta);
+        if (cl.getLevelMetaData(playtestname, meta)) {
+            cl.ListOfMetaData.clear();
+            cl.ListOfMetaData.push_back(meta);
         } else {
-            ed.loadZips();
-            if (ed.getLevelMetaData(playtestname, meta)) {
-                ed.ListOfMetaData.clear();
-                ed.ListOfMetaData.push_back(meta);
+            cl.loadZips();
+            if (cl.getLevelMetaData(playtestname, meta)) {
+                cl.ListOfMetaData.clear();
+                cl.ListOfMetaData.push_back(meta);
             } else {
                 vlog_error("Level not found");
                 VVV_exit(1);
@@ -635,8 +638,8 @@ int main(int argc, char *argv[])
         }
 
         game.loadcustomlevelstats();
-        game.customleveltitle=ed.ListOfMetaData[game.playcustomlevel].title;
-        game.customlevelfilename=ed.ListOfMetaData[game.playcustomlevel].filename;
+        game.customleveltitle=cl.ListOfMetaData[game.playcustomlevel].title;
+        game.customlevelfilename=cl.ListOfMetaData[game.playcustomlevel].filename;
         if (savefileplaytest) {
             game.playx = savex;
             game.playy = savey;
